+++ keycard/commands/generate_menmonic.py
@@ -4,16 +4,21 @@ from ..preconditions import require_secure_channel
 
 
 @require_secure_channel
-def generate_mnemonic(card: CardInterface, checksum_size: int = 6) -> list[int]:
+def generate_mnemonic(
+    card: CardInterface,
+    checksum_size: int = 6
+) -> list[int]:
     """
     Generate a BIP39 mnemonic using the card's RNG.
 
     Args:
         card (CardInterface): The card interface.
-        checksum_size (int): Number of checksum bits (between 4 and 8 inclusive).
+        checksum_size (int): Number of checksum bits
+            (between 4 and 8 inclusive).
 
     Returns:
-        List[int]: List of integers (0-2047) corresponding to wordlist indexes.
+        List[int]: List of integers (0-2047) corresponding to wordlist
+            indexes.
 
     Raises:
         ValueError: If checksum size is outside the allowed range.